backend/chatbot/requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
requests